Massive Pull-Request!
I've made a LOT of changes in this pull request but I can assure you that I have checked and these changes are real. The 1 thing I'm not sure about is the Artist rank. I heard about it when you were talking about it in stream so if it's true, I've pre-made it for you! :D
(I've removed /getadmin as it's to get Event Admin or MPS Admin on an Arcade server, not for admin perms in the lobby. It used to be like that but it's been changed)